@dongjoon-hyun 2025-01-05
Hi, @LuciferYang . It seems that this PR causes massive failures on Apple Silicon CIs which uses this method.

- https://github.com/apache/spark/actions/workflows/build_maven_java21_macos15.yml
  - https://github.com/apache/spark/actions/runs/12616477101
  - https://github.com/apache/spark/actions/runs/12614014000

Previously, there was a few flaky test pipelines, but now most of test pipelines fails. Could you double check the CI?
<img width="304" alt="Screenshot 2025-01-05 at 10 33 15" src="https://github.com/user-attachments/assets/fe42af44-9546-4200-ba04-5b92c5cc627b" />