@LuciferYang 2025-01-05
Thanks for your report, I will take a look at this issue during today's working hours.